+++ libcnb-data/src/buildpack/mod.rs
@@ -92,7 +92,7 @@ pub enum BuildpackDescriptor<BM> {
 /// assert_eq!(buildpack_descriptor.buildpack.id, buildpack_id!("foo/bar"));
 /// assert_eq!(buildpack_descriptor.stacks, vec![Stack::Any]);
 /// ```
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Debug, Clone)]
 #[serde(deny_unknown_fields)]
 pub struct SingleBuildpackDescriptor<BM> {
     pub api: BuildpackApi,
@@ -147,7 +147,7 @@ pub struct MetaBuildpackDescriptor<BM> {
     pub metadata: BM,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Debug, Clone)]
 #[serde(deny_unknown_fields)]
 pub struct Buildpack {
     pub id: BuildpackId,
@@ -169,7 +169,7 @@ pub struct Buildpack {
     pub sbom_formats: HashSet<SbomFormat>,
 }
 
-#[derive(Deserialize, Debug, Eq, PartialEq)]
+#[derive(Deserialize, Debug, Eq, PartialEq, Clone)]
 #[serde(deny_unknown_fields)]
 pub struct License {
     pub r#type: Option<String>,