+++ libcnb-data/src/lib.rs
@@ -14,6 +14,7 @@ pub mod layer;
 pub mod layer_content_metadata;
 pub mod sbom;
 pub mod store;
+pub mod buildpackage;
 
 mod newtypes;