+++ libcnb-package/src/lib.rs
@@ -10,13 +10,14 @@ pub mod cross_compile;
 
 use crate::build::BuildpackBinaries;
 use libcnb_data::buildpack::SingleBuildpackDescriptor;
+use libcnb_data::buildpackage::Buildpackage;
 use std::fs;
 use std::path::{Path, PathBuf};
 
 /// The profile to use when invoking Cargo.
 ///
 /// <https://doc.rust-lang.org/cargo/reference/profiles.html>
-#[derive(Copy, Clone)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum CargoProfile {
     /// Provides faster compilation times at the expense of runtime performance and binary size.
     Dev,
@@ -45,15 +46,44 @@ pub fn read_buildpack_data(
         })
 }
 
+pub fn read_buildpackage_data(
+    project_path: impl AsRef<Path>,
+) -> Result<BuildpackageData, BuildpackageDataError> {
+    let buildpackage_descriptor_path = project_path.as_ref().join("package.toml");
+
+    fs::read_to_string(&buildpackage_descriptor_path)
+        .map_err(BuildpackageDataError::IoError)
+        .and_then(|file_contents| {
+            toml::from_str(&file_contents).map_err(BuildpackageDataError::DeserializationError)
+        })
+        .map(|buildpackage_descriptor| BuildpackageData {
+            buildpackage_descriptor_path,
+            buildpackage_descriptor,
+        })
+}
+
 #[derive(Debug)]
 pub enum BuildpackDataError {
     IoError(std::io::Error),
     DeserializationError(toml::de::Error),
 }
 
+#[derive(Debug, Clone)]
 pub struct BuildpackData<BM> {
     pub buildpack_descriptor_path: PathBuf,
     pub buildpack_descriptor: SingleBuildpackDescriptor<BM>,
+}
+
+#[derive(Debug)]
+pub enum BuildpackageDataError {
+    IoError(std::io::Error),
+    DeserializationError(toml::de::Error),
+}
+
+#[derive(Debug)]
+pub struct BuildpackageData {
+    pub buildpackage_descriptor_path: PathBuf,
+    pub buildpackage_descriptor: Buildpackage,
 }
 
 /// Creates a buildpack directory and copies all buildpack assets to it.