+++ CHANGELOG.md
@@ -5,6 +5,7 @@ separate changelogs for each crate were used. If you need to refer to these old 
 `HISTORICAL_CHANGELOG.md` in their respective crate directories.
 
 ## [Unreleased]
+- Changes to `libcnb package` command to support compiling buildpacks and meta-buildpacks from projects containing a single buildpack or multiple buildpacks in a monorepo.
 
 ## [0.11.5] 2023-02-07